+++ src/app/layout.tsx
@@ -2,14 +2,12 @@ import './globals.css';
 
 import GnbRenderer from '@common/gnb/GnbRenderer'; // dynamic 없이 일반 import
 import ModalManager from '@common/modal/ModalManager';
+import { Metadata, Viewport } from 'next';
 import localFont from 'next/font/local';
 import { SessionProvider } from 'next-auth/react';
 import { ThemeProvider } from 'next-themes';
 
 import { Providers } from '@/app/providers';
-
-import { metadata } from './metadata';
-import { viewport } from './viewport';
 
 const pretendard = localFont({
   src: '../../public/fonts/PretendardVariable.woff2',
@@ -18,18 +16,24 @@ const pretendard = localFont({
   weight: '100 900',
 });
 
-// eslint-disable-next-line react-refresh/only-export-components
-export { metadata, viewport };
+export const metadata: Metadata = {
+  title: 'Albamate',
+  description: 'Albamate application', // 추후 변경 가능
+  authors: [{ name: 'Albamate Team' }],
+  keywords: ['알바', '채용', '구인', '구직', 'Albamate'],
+};
+
+export const viewport: Viewport = {
+  themeColor: '#ffffff',
+  initialScale: 1,
+  maximumScale: 1,
+  minimumScale: 1,
+  userScalable: false,
+};
 
 const RootLayout = ({ children }: { children: React.ReactNode }) => {
   return (
     <html suppressHydrationWarning lang="ko">
-      <head>
-        <meta
-          content="width=device-width, initial-scale=1, maximum-scale=1, minimum-scale=1, user-scalable=no"
-          name="viewport"
-        />
-      </head>
       <body className={`${pretendard.className} dark:bg-gray-900`}>
         <SessionProvider>
           <ThemeProvider enableSystem attribute="class" defaultTheme="system">
