--- src/app/metadata.ts
@@ -1,8 +1,0 @@
-import type { Metadata } from 'next';
-
-export const metadata: Metadata = {
-  title: 'Albamate',
-  description: 'Albamate application', // 추후 변경 가능
-  authors: [{ name: 'Albamate Team' }],
-  keywords: ['알바', '채용', '구인', '구직', 'Albamate'],
-};
