--- src/app/viewport.ts
@@ -1,5 +1,0 @@
-import type { Viewport } from 'next';
-
-export const viewport: Viewport = {
-  themeColor: '#ffffff',
-};
